+++ tests/server-test.js
@@ -101,7 +101,7 @@ describe('src/server', function () {
         });
 
         sinon.assert.callCount(instrumenter.countHit, 1);
-        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'rule', '');
+        sinon.assert.calledWith(instrumenter.countHit, 'accepted', '');
 
         sinon.assert.callCount(instrumenter.timeHit, 1);
         // Since we've installed sinon's fake timers we can safely compare to new Date()
@@ -140,7 +140,7 @@ describe('src/server', function () {
         });
 
         sinon.assert.callCount(instrumenter.countHit, 1);
-        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'rule', 'get-ping-by-ip');
+        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'get-ping-by-ip');
 
         sinon.assert.callCount(instrumenter.timeHit, 1);
         // Since we've installed sinon's fake timers we can safely compare to new Date()
@@ -175,7 +175,7 @@ describe('src/server', function () {
         });
 
         sinon.assert.callCount(instrumenter.countHit, 1);
-        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'rule', '');
+        sinon.assert.calledWith(instrumenter.countHit, 'accepted', '');
 
         sinon.assert.callCount(instrumenter.timeHit, 1);
         // Since we've installed sinon's fake timers we can safely compare to new Date()
@@ -184,6 +184,50 @@ describe('src/server', function () {
 
         done();
       }).catch(done);
+    });
+
+    it('for an operation that matches a passing canary rule', function () {
+      backend.hit.returns(Promise.resolve({
+        isAllowed: true,
+        currentCredit: 100,
+        nextResetSeconds: 60,
+      }));
+      return client.hit({
+        method: 'GET',
+        path: '/ping',
+        local: 'true',
+        ip: '1.2.3.4',
+      }).then(() => {
+        sinon.assert.callCount(instrumenter.countHit, 2);
+        sinon.assert.calledWith(instrumenter.countHit, 'canary-accepted', 'get-ping-by-ip-from-local');
+        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'get-ping-by-ip');
+        sinon.assert.callCount(instrumenter.timeHit, 1);
+      });
+    });
+
+    it('for an operation that matches a rejecting canary rule rule', function () {
+      backend.hit.onCall(0).returns(Promise.resolve({
+        isAllowed: false,
+        currentCredit: 0,
+        nextResetSeconds: 60,
+      }));
+      backend.hit.onCall(1).returns(Promise.resolve({
+        isAllowed: true,
+        currentCredit: 100,
+        nextResetSeconds: 60,
+      }));
+
+      return client.hit({
+        method: 'GET',
+        path: '/ping',
+        local: 'true',
+        ip: '1.2.3.4',
+      }).then(() => {
+        sinon.assert.callCount(instrumenter.countHit, 2);
+        sinon.assert.calledWith(instrumenter.countHit, 'canary-rejected', 'get-ping-by-ip-from-local');
+        sinon.assert.calledWith(instrumenter.countHit, 'accepted', 'get-ping-by-ip');
+        sinon.assert.callCount(instrumenter.timeHit, 1);
+      });
     });
 
     it('for an unknown command', function (done) {